Postgres persistence behind one environment variable, with SQLite retained
QUINTEK_DATABASE_URL unset keeps SQLite and no install step; set selects
PostgreSQL from a bounded pool. That is the whole switch, which is what makes
the rollback real -- a bad deployment is reverted by clearing a variable, not
by reverting code.

The audit's four incompatibilities are fixed, each with a regression test
proven to fail on the unported code:

  1. source_concepts and gap_links put a NULLABLE column inside a composite
     PRIMARY KEY. SQLite permitted it and then accepted duplicate rows, so
     INSERT OR IGNORE never deduplicated -- a live SQLite bug, not only a
     Postgres one. Postgres silently promotes the column to NOT NULL and
     rejects the insert on real data. Replaced with partial unique indexes,
     which express the intent correctly on both.
  2. BEGIN IMMEDIATE has no Postgres equivalent and the obvious translation
     reopens the allowance-overspend bug it exists to prevent: measured, two
     200-unit requests authorised 400 against a 300 cap. Replaced with a
     transaction-scoped advisory lock keyed per user, which is stricter where
     it matters and looser where it does not -- two learners no longer
     serialise behind each other as they did under SQLite's database-wide
     write lock.
  3. revision.py grouped by c.id while selecting concept_state columns.
     Rejected outright by Postgres; now grouped fully, which is exactly
     equivalent because the join pins concept_state's whole primary key.
  4. cost_ledger's micro-unit columns are BIGINT. A USD 0.30/M model already
     in configs/model_prices.json converts to 2,550,000,000 micro-paise, past
     the 32-bit ceiling.

Two further defects surfaced while testing rather than by inspection:

  * A caught duplicate webhook left the transaction aborted and handed a
    poisoned connection back to the pool. Postgres connections now run in
    autocommit, matching billing/db.py's existing SQLite isolation_level=None,
    so a standalone failure does not poison the session while a failure inside
    an explicit transaction still aborts it.
  * Re-running schema.sql per connection deadlocked two threads on pg_proc.
    Initialisation is now once per process and guarded by an advisory lock so
    two instances booting together serialise instead of racing.

Connections are returned at the end of each request. ThreadingHTTPServer
starts a thread per request and the caches are thread-local, so without that
the ninth concurrent request against a pool of eight blocks while the service
still looks healthy.

student/db.py's docstring claimed nothing below used a SQLite-only feature.
That was the sentence the migration decision leaned on, and it was wrong; it
now records what the audit actually found.

Tests: 1371 passed / 25 skipped on SQLite alone; 1392 passed / 4 skipped with
QUINTEK_TEST_POSTGRES_URL set against PostgreSQL 16. Baseline was 1353/4.
Postgres tests SKIP without that variable and are never reported as passing.

+++ billing/schema.sql
@@ -180,8 +180,16 @@ CREATE TABLE IF NOT EXISTS cost_ledger (
     cached_tokens        INTEGER,
     -- Provider-reported price per million tokens, in micro-minor-units so a
     -- fraction of a paise survives without a float.
-    price_in_micro       INTEGER,
-    price_out_micro      INTEGER,
+    --
+    -- BIGINT, NOT INTEGER, and this is not precautionary. A price of USD 0.30
+    -- per million output tokens -- a CHEAP model, and one in
+    -- configs/model_prices.json today -- converts to
+    -- 0.30 * 8500 paise/USD * 1_000_000 = 2,550,000,000 micro-paise, which is
+    -- past the 2,147,483,647 ceiling of a 32-bit integer. SQLite never
+    -- complained because its INTEGER is 64-bit whatever the column says;
+    -- PostgreSQL's INTEGER is 32-bit and rejects the row outright.
+    price_in_micro       BIGINT,
+    price_out_micro      BIGINT,
     -- MICRO minor units (millionths of a paise), not paise.
     --
     -- A single generation call costs a fraction of a paise. Storing that as
@@ -189,7 +197,13 @@ CREATE TABLE IF NOT EXISTS cost_ledger (
     -- row inflated a measured 10,000-call total from Rs 30 to Rs 100 -- a
     -- 3.3x over-report of the single line the economics dashboard exists to
     -- get right. Precision is kept here and rounded ONCE at aggregation.
-    cost_micro           INTEGER NOT NULL DEFAULT 0,
+    --
+    -- BIGINT for the same reason as the price columns above. A single call's
+    -- cost is far smaller than the per-million rate, so this one is headroom
+    -- rather than an observed overflow -- but SUM() over it is the number the
+    -- economics dashboard reports, and a 32-bit column is a poor place to
+    -- discover that a month of inference does not fit.
+    cost_micro           BIGINT NOT NULL DEFAULT 0,
     currency             TEXT NOT NULL DEFAULT 'INR',
     compute_units        INTEGER NOT NULL DEFAULT 0,
     questions_produced   INTEGER NOT NULL DEFAULT 0,
